# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Erudit API
  version: 1.0.0
  description: Контракт всех реализованных и запланированных REST-эндпоинтов. Операции с пометкой «Запланировано»
    появятся после выполнения соответствующих задач; Try it out для них пока вернёт 404.
servers:
- url: /
...
paths:
...
  /api/v1/admin/content/moderation:
    get:
      tags:
      - ContentModeration
      operationId: listModerationQueue
      summary: Очередь модерации
      description: Возвращает ожидающий решения контент. Требуется роль ADMIN.
      x-task-id: 20
      x-implementation-status: implemented
      parameters: *id004
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentListResponse'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-required-role: ADMIN
  /api/v1/admin/content/{id}/approve:
    post:
      tags:
      - ContentModeration
      operationId: approveContent
      summary: Одобрить контент
      description: Публикует контент из очереди модерации. Требуется роль ADMIN.
      x-task-id: 20
      x-implementation-status: implemented
      parameters:
      - name: id
        in: path
        required: true
        description: Идентификатор ресурса
        schema: *id001
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentItemResponse'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-required-role: ADMIN
  /api/v1/admin/content/{id}/reject:
    post:
      tags:
      - ContentModeration
      operationId: rejectContent
      summary: Отклонить контент
      description: Отклоняет контент из очереди модерации. Требуется роль ADMIN.
      x-task-id: 20
      x-implementation-status: implemented
      parameters:
      - name: id
        in: path
        required: true
        description: Идентификатор ресурса
        schema: *id001
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentItemResponse'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-required-role: ADMIN
  /api/v1/admin/content/reports:
    get:
      tags:
      - ContentModeration
      operationId: listContentReports
      summary: Жалобы на контент
      description: Запланировано в задаче 61. До реализации этот маршрут вернёт 404. Требуется роль ADMIN.
      x-task-id: 61
      x-implementation-status: planned
      parameters: *id004
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentReportListResponse'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-required-role: ADMIN
  /api/v1/admin/content/reports/{id}/resolve:
    post:
      tags:
      - ContentModeration
      operationId: resolveContentReport
      summary: Рассмотреть жалобу
      description: Запланировано в задаче 61. До реализации этот маршрут вернёт 404. Требуется роль ADMIN.
      x-task-id: 61
      x-implementation-status: planned
      parameters:
      - name: id
        in: path
        required: true
        description: Идентификатор ресурса
        schema: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportResolutionRequest'
      responses:
        '204':
          description: Операция выполнена
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-required-role: ADMIN
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
      - code
      - message
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
      required:
      - error
    PageMetadata:
      type: object
      properties:
        page: &id007
          type: integer
          format: int32
        size: *id007
        totalElements:
          type: integer
          format: int64
        totalPages: *id007
      required:
      - page
      - size
      - totalElements
      - totalPages
...
    ContentItem:
      type: object
      properties:
        id: *id001
        categoryId: *id001
        type:
          type: string
          enum: [ARTICLE, FACT_CARD, AUDIO, VIDEO, COMBINED, QUIZ_BLOCK]
        title:
          type: string
          minLength: 1
          maxLength: 250
        description:
          type: string
          maxLength: 2000
        body:
          type: string
        mediaUrl:
          type: string
        difficulty:
          type: string
        estimatedMinutes: *id007
        authorId:
          type: string
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        premiumLocked: *id008
      required:
      - id
      - type
      - title
      - status
...
    ContentReportItem:
      type: object
      properties:
        id: *id001
        contentId: *id001
        reason:
          type: string
        status:
          type: string
        createdAt: *id005
      required:
      - id
      - contentId
      - status
...
    ReportResolutionRequest:
      type: object
      properties:
        decision:
          type: string
          enum:
          - ACCEPT
          - REJECT
        comment:
          type: string
      required:
      - decision
...
    ContentItemResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ContentItem'
      required:
      - data
    ContentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ContentItem'
        pagination:
          $ref: '#/components/schemas/PageMetadata'
      required:
      - data
...
    ContentReportListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ContentReportItem'
        pagination:
          $ref: '#/components/schemas/PageMetadata'
      required:
      - data
...
security:
- bearerAuth: []